samples: wifi: ble_coex: Fix crash when Zperf is started
Zperf parameters data structure is not initialized, and with the upmerge
changes a function pointer was added as part of the parameters, and if
the uninitialized struct has set this to non-zero, then a random
address is called by the custom loader function pointer, this causes a
crash.

Initialize the parameters to zero.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>