tools: Upgrade `nrfutil` from 8.0.0 to 8.1.0
Including this version instead of version 8.0.0 would be preferred as it
supports the "locked" mode in `nrfutil`. This mode, which is controlled
by inserting a file named "locked" in the root of NRFUTIL_HOME, allows
locking all functionality that changes the installed versions, both of
`nrfutil` itself and of installed command. This will lower the risk of
users accidentally upgrading the bundled `nrfutil`/`nrfutil device`.

Signed-off-by: Herman Skogseth <herman.skogseth@me.com>