samples|tests: drivers: Adjust overlays to align with sQSPI v1.0.0
Offsets of the register interface for sQSPI soft peripherals (so in
consequence the base addresses of the related nodes in dts) have been
changed in sQSPI v1.0.0 for both the nRF54l15 and nRF54h20.
Update accordingly dts overlays in samples and tests, and also the sQSPI
MSPI shim driver documentation and its cmake definitions.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>