nrf_security: Allow setting the kmu reserved area with dts
The kmu_push_area was only allowed to be set with dts
for the ram images. Now more images will use it and
hence the dependency on RAM images is removed.

Also create a Kconfig which depends on the dts entry
so that it can be used in the build system.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>