modules: openthread: Fix OpenThread CSL Accuracy
The current accuracy is set to 50ppm, which is not compliant
with IEEE 802.15.4-2015. Moreover, we cannot use the
CONFIG_CLOCK_CONTROL_NRF_ACCURACY kconfig as a return in the
otPlatRadioGetCslAccuracy function because it must be
configurable by users, depending on the used oscillators.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>
(cherry picked from commit aec518c08dcb49dab794a50a725962b77f77bb56)