nrf_security: IronSide SE support for PSA ITS API
Allow IronSide SE to export the `psa_its_*` API over IPC. This makes it
so that the Internal Trusted Storage can be used to store not only keys
(via the PSA Crypto API) but other sensitive assets too.

Once again, the serialization code is reused from TF-M, which leverages
the IronSide SE implementation of `psa_call()` that was initially added
to support the PSA Crypto API.

The only change to `psa_call()` is that the `type` argument is no longer
ignored. It has to be sent over IPC, because it identifies the ITS API.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>