samples: zephyr: smp_svr_mini_boot: Add 'swap using move' config
This commit adds a new sample configuration for the SMP server
mini boot sample that uses the swap using move mode of MCUboot.
This configuration is useful for testing and demonstrating the
functionality of the swap using move feature in a minimal setup.

Documentation has been updated with instructions on how to build
the sample with direct-xip mode or swap using move mode.

Signed-off-by: Michal Kozikowski <michal.kozikowski@nordicsemi.no>