doc: matter: Add known issue entry for PSA NOC bug
If a Matter controller attempts to commission a device to the same
Matter fabric to which it is already commissioned, the failsafe
procedure is triggered.

In this case, the new NOC key is discarded, but the old NOC key
is not restored.
This issue is caused by a bug where the same PSA key ID is used
for both the active and pending NOC keys.
When the pending key is being destroyed, the active key is also
destroyed. As a result, the TC-OPCREDS-3.8 certification test
fails.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>