modules: tf-m: Change default profile for most use cases
Change the default profile for TF-M for all devices which
use the bluetooth crypto, bluetooth mesh crypto and openthread.

Instead of the minimal profile of TF-M which doesn't provide
crypto operations apart from rng use the PROFILE_TYPE_NOT_SET
which allows crypto operations based on the enabled PSA_WANTs.

This is preparation work, the profile will be PROFILE_TYPE_NOT_SET
for all devices and use cases in the future.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>