applications: nrf5340_audio: Volume 3dB low for BIS-16_2_1 to 48k resampler
OCT-3428

Volume for the BIS 1_2_16 level is significantly lower
compared to the default build without resampling. The
difference is rather precisely -3dB for this configuration,
which is equal to one entire volume-down button press.

This was tested using a 1kHz sine wave tone and I2S as
audio source pathway, and the audio level was measured
at the analog output of the audio sink. As such, the
discrepancy can be either a misconfiguration of the
hardware codec, or a bug in the resampler code.

Signed-off-by: Graham Wacey <graham.wacey@nordicsemi.no>