samples: bluetooth: Add period_in_events to conn event trigger sample
As well as conn_evt_counter_start.

This shows a simple way of setting up a GPIO on 54L15 with this feature
such that it toggles every N connection events from connection
event counter A. In this case every 2 events from event #150. It's
one of several simple approaches.

This makes setup_connection_event_trigger() fully compatible with the
old SoftDevice API with a similar name, except this one works on the
new chips as well.

A single link is assumed. With multiple links/activities the connection
can get blocked so some extra logic would be needed.

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>