dfu_multi_image: Add the possibility to save progress
This commit adds the possibility to restore progress
when performing an update using dfu_multi_image
in case of a reset.

Also, generation of dfu_multi_image now adds padding
add the end of the dfu_multi_image header and each
of the images (except for the last one, where it is
not necessary) so that all of the images start at
a given alignment. The alignment is configured  by
the CONFIG_DFU_MULTI_IMAGE_ALIGN option and should
be set to the write alignment of the dfu_targets
memory.
This is necessary when performing un-buffered writes
(which is the case when saving the progress). The
chunks in which the dfu_multi_image package is
sent are aligned, so an unaligned image inside
the package would result in unaligned writes
of one of the subsequent chunks and in turn
in write errors.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>