Revert "samples: bluetooth: switch from hci_ipc to ipc_radio in sample tests"
This reverts commit 90fc84672db08806d7951c67a09ef4393cf0d2eb.

The reason for reverting is poor performance of the ipc_radio driver.
For example, l2cap events that used to take fraction of a second,
now takes roughly 10 times longer, i.e. 6 seconds.

This revert is a temporary measure until the performance issues
are further investigated and resolved.
The switch to the IPC Radio driver will be re-evaluated
once a fix is implemented.

Signed-off-by: Aleksandr Mirlenko <aleksandr.mirlenko@nordicsemi.no>