Revert "samples: bluetooth: switch from hci_ipc to ipc_radio in cs samples"
This reverts commit 40cca60f6906f98bd23cd344197892b3f4aadbb8.

The reason for reverting is poor performance of the ipc_radio driver.
For example, l2cap events that used to take fraction of a second,
now takes roughly 10 times longer, i.e. 6 seconds.

This revert is a temporary measure until the performance issues
are further investigated and resolved.
The switch to the IPC Radio driver will be re-evaluated
once a fix is implemented.

Signed-off-by: Aleksandr Mirlenko <aleksandr.mirlenko@nordicsemi.no>