west_commands: ncs_ironside_se_update: fix various issues
Only use the --x-sdfw-variant option when it is supported by nrfutil,
since it is not available in all versions of nrfutil-device.
The program helptext is used to determine if it is supported.

Add an explicit 'nrfutil device recover' ahead of the update,
which seems to be necessary in some cases.

Fix an issue where the wrong update status was logged after
the update of a slot.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>
(cherry picked from commit 5204e99b192bc972be038cefbc29bb9691425c9f)