nrf_security: Use static variables in sx_blkcipher_ecb_simple
The in and out dma descriptors variables in sx_blkcipher_ecb_simple
takes up around 80 bytes of space in the stack.
To optimize the stack usage these variables are now static, the
mutex locking was moved before there variables are set so that
accessing these from multiple threads is not an issue.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>