nrf_security: Refactor blkcipher in Cracen
In the Cracen driver the PSA context cracen_cipher_operation_t
already contains a struct sxkeyref and also a struct sxblkcipher
which also contains a sxkeyref struct. Instead of keeping two
instanses of the sxkeyref keep the one in the PSA operation
and convert the other one to pointer and refactor the code
of blkcipher accordingly.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>