lib: lte_link_control: Fix missing eDRX events with non-NTN firmware
The library tried to always configure eDRX also for NTN NB-IoT. This
causes an error with non-NTN modem firmware, which also causes the
+CEDRXS notification subscription to be removed. To mitigate this,
the order in which eDRX is configured for access technologies has
been changed. This commit also adds checking of the modem firmware
type, so that configuring eDRX for NTN NB-IoT is skipped when
non-NTN firmware is detected.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>