applications: nrf_desktop: nrf54h20: migrate to ironside se
Updated the nrf54h20dk/nrf54h20/cpuapp and nrf54h20dk/nrf54h20/cpurad
board target configurations in the nRF Desktop application to align
them with the new nRF54H20 SW architecture that is based on the
IronSide SE.

Added the MCUboot bootloader configuration and replaced the default
nRF54H20 partition map with the custom application-optimized memory
map.

The updated configuration is the first step towards the complete
feature migration. The following features are disabled and not yet
supported:

- MCUmgr DFU with the Bluetooth transport
- Config Channel DFU with both the USB and Bluetooth transports
- DVFS
- S2RAM in dongle configurations

Ref: NCSDK-34151

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>