modules: tf-m: Add Kconfig option for older PS format
TF-M recently added a new format for the protected storage data.
In order to keep backwards compatibility it added the configuration
TFM_PS_SUPPORT_FORMAT_TRANSITION to allow new versions of TF-M to read
the older data format.

This allows the option to be selected as a Kconfig option
in sdk-nrf.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>