modules: tf-m: Refactor cmake logic
Refactor the cmake logic for a couple of options
in TF-M.

This remove the generator expressions for the mcuboot
single image option and adds them in an existing if
statement as they used to be before the upmerge.

Do the same for the hw init nrf peripherals option as well.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>