memfault: upgrade to 1.29.0
Upgrade the Memfault Firmware SDK to version 1.29.0 to include a fix
required to remove a Kconfig style issue. The redefinition includes a
`default n`, which causes the Kconfig style check shared in
https://github.com/zephyrproject-rtos/zephyr/issues/94780#issuecomment-3210257794 to fail.
Kconfig style check script passes after this change.

Additionally, confirmed that the Kconfig values are expected from
building the samples/debug/memfault app for a nrf9160 target:
- CONFIG_NRF_MODEM_LIB=y
- CONFIG_MEMFAULT_HTTP_USES_MBEDTLS is not set

Note this is a follow-on PR to https://github.com/nrfconnect/sdk-nrf/pull/24402.

Signed-off-by: Gillian Minnehan <gillian.minnehan@nordicsemi.no>