bluetooth: cs_de: Fix after combined main and sub mode CS steps
There was a change in sdk-zephyr Bluetooth host implementation
that combined main and sub CS mode steps.

Not all combinations of main mode and sub mode are valid, so to make it
harder to select incorrect parameters, the main mode and sub mode
enums are combined into one.

With that change names of some CS related enums were changed.
The commit aligns sdk-nrf to use the new names.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>