bluetooth: make CONFIG_BT_RX_STACK_SIZE adjustment common for 54LX SoCs
The Bluetooth stack requires larger BT_RX_STACK_SIZE for nRF54LX SoCs
family and nRF54H20 due to higher memory requirements for crypto on
these SoCs.

The Kconfig was modified in couple of samples. The change was added
for next SoCs based on these two platforms. Instead of that it should
be a generic configuration is a target is nRF54LX or nRF54HX.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>