bluetooth: cs_de: Less strict reqs to accept measurements as OK quality
Updates the requirements for how we determine if a ranging has enough
good tones to calculate distance estimates.
From requiring no more than 4 channels with "bad" tones
to requiring at least 15 channels with "good" tones.
Note that CS requires at least 15 channels to be enabled.

The reason for making this change is that the requirement of no more
than 4 channels with "bad" tones becomes quite hard to satisfy at
ranges over a few meters, especially in noisy environements,
which would prevcent cs_de from producing new distance estimates.
By changing the requirement we can produce new distance estimates more
regularly, even if the estimates may be less accurate.
The reasoning here is that even if the estimates are
less accurate it is better to have regular updates, and filters can
be applied on top of the estimates produced to filter out outliers/bad
estimates.

Signed-off-by: Erik Sandgren <erik.sandgren@nordicsemi.no>