applications: nrf_desktop: nrf54h20: disable power domains in MCUboot
Disabled the power domain Kconfig option for the
nrf54h20dk/nrf54h20/cpuapp board target in the MCUboot image
configuration of the nRF Desktop application. This Kconfig option
is not needed in the bootloader firmware.

Disabled relevant DTS configuration that is also redundant when
the power domain configuration is disabled.

Ref: NCSDK-35587

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>