tests: power_consumption: common: use main thread instead of timer
The main thread is a thread, and thus has a larger stack and the
ability to block which is required for controlling GPIOs on the
nRF54H20. The main thread has a default prio of 0, the thread
used by the tests has a prio of 5, so they will be preempted and
suspended by the main thread once it times out. Timers run in
interrupt context which is what caused the issue for the nrf54h20.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>