sysbuild: nrf54l: Add protected ram invalidation logic
This adds the logic that is needed for nRF54L devices in order
to provision the protected ram invalidation slots using nrfutil.

This automatically generates a json file with the appropriate
PSA attributes which is provisioned every time that the device
is flashed.

Until now this provisioning happens in the application which uses
non-volatile and volatile memory. With this solution the application
doesn't need to do that manually anymore.

Also minor but rename rotable->rotatable because according to
the Merriam Webster dictionary rotable is not a word.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>