scripts: hid_configurator: remove SUIT support
Removed the SUIT support from the HID configurator script due to the
general migration to the IronSide SE solution for the nRF54H20 platform.
As part of the IronSide SE transition, the DFU solution based on the
SUIT has been first deprecated and then removed from the nRF Connect
SDK.

This change aligns the script with the remaining parts of the nRF
Connect SDK. The nRF Desktop application now supports the MCUboot-based
DFU for the nRF54H20 DK target.

Ref: NCSDK-34151

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>