boot: Adjustments to region protection code for all nRF54L platforms
This commit fixes a couple of issues regarding B0 and
MCUBoot region protection for nRF54L. It also adds some tests
verifying that the issues are no longer present.

Also, support for region and BOOTCONF protection is added for
nRF54LM20 and nRF54LV10 platforms.

Note: due nRF54LM20 and nRF54LV10 devices being shipped
in the TEST mode, the BOOTCONF configuration is not copied
to the appropriate REGION[n].CONFIG register and is not
automatically applied at startup.
Thus, the BOOTCONF configuration does not work with the
shipped devices. However, all the code needed for
that is present.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>