nrf_security: Allow importing the nRF54L PROT_RAM invalidation slots
Add a Kconfig option to allow performing a psa_import_key in order
to provision the KMU slots used for invalidating the protected RAM
content.

With this change a psa_import_key with KMU slot 248 will populate
the KMU invilidation slots with random data so that it can be used
later.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>