lib: Add OTDOA Adaption Layer library
Adds a new library supporting the hellaPHY OTDOA
cellular positioning method.

hellaPHY OTDOA is a positioning method that uses
cellular signals to estimate a UE position.  A library
that implements it is distributed separately from the
nRF Connect SDK.

The adaption layer library is a source code library intended
to be distributed as a part of the nRF Connect SDK.  It is
used to integrate the hellaPHY OTDOA library with an
nRF Connect SDK Application

Signed-off-by: Neal Riedel <nriedel@serranosystems.com>