tests: benchmarks: current_consumption: rram_write: Fix address check
Test writes data to the TEST_AREA in chunks of BUF_SIZE bytes.
Currently, test checks whether the beginning of the chunk
is within the TEST_AREA. This does not guarantee that the last
chunk fits completely.

Fix address checking. Check whether the entire chunk fits
within the TEST_AREA.

Signed-off-by: Sebastian Głąb <sebastian.glab@nordicsemi.no>