samples: dfu: Add single slot sample
Added a sample with a very simple main application
and minimized firmware loader image and MCUBoot.
It presents how to maximize the space available to
the application.

The firmware loader in the sample is entered by pressing
a button during reset.
The functionality of the sample is pretty similar to
the firmware_loader_entrance sample, but it is simplified.
This will also allow much easier presentation of how
to add firmware loader support for nRF54H20 in the future
(currently the sample only supports nRF54L15).

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>