nrf_modem_lib: add support for SO_SENDCB
Added support for SO_SENDCB in the socket offloading layer. This socket
option supported by the modem firmware mfw_nrf9151-ntn allows the user
to set a callback function that will be called when a send operation is
done sending data on air.
In order to map the nRF socket descriptor received from the modem to the
appropriate ZVFS descriptor, the latter is now also stored in the
offload socket context.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>