modules: tf-m: do not include MDK/nrfx in RTE_Device.h
This file gets included by nrfx_config.h in TF-M for RTE defines.

With the update of nrfx to 4.0 includes of nrf.h (MDK) are replaced by
nrfx.h (nrfx), so we end up in a chicken-and-egg situation where the
configuration header file for nrfx includes and depends on nrfx itself.

To remedy that, replace uses of MDK defines by TF-M compile definitions
and remove the header file inclusion.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>