doc: Improve and expand the IronSide SE docs
Primarily, we are now documenting that the user can configure UICR via
Kconfig. But there are also minor improvements and corrections related
to IronSide SE all over the place.

This is just a step in the correct direction, and there are still many
improvements to be made.

To make the doc changes reviewable and to reduce git conflicts I am
improving the documentation in steps.

One notable thing left as future-work:

There is a lot of good information about how to configure UICR in the
Kconfig help text. But referencing this information from .rst is
blocked on

https://github.com/zephyrproject-rtos/zephyr/pull/98911

I have changed some of the references to 'UICR.<FEAURE>' to be
references to Kconfig options. In the hope that after 98911 these
could be links to the Kconfig help text.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>