lib: lte_lc: Integrate PDN library
Copies the PDN functionality from the PDN library that is about
to be deprecated, and adds it as a module in the LTE link
controller library.

The functionality is the same, but with naming changes to types
and APIs to align with the rest of LTE link controller.
In addition, PDN events are now part of the LTE LC events, and
there is no longer need for a separate event handler.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>