tests: drivers: lpuart: Extend and fix interrupt driven test suite
In resiliency test for the interrupt driven API wrong approach was used.
When TX is used then REQ line cannot float because that is the line the
TX controls. Reworked floating pin test helper to support toggling REQ
or RDY pin. Added test case where DUT attemts to TX and test simulates
that receiver in not lpuart and pin floats or changes state. Reworked
existing test so that REQ pin floats only when receiver is opened and
TX is not active.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>