bluetooth: cs_de: Set RTT estimate to 0.0 if ToF is negative
It might happen that the ToF is negative because the RTT measurements
are in CS are not perfectly accurate and may have a statisical
distribution around its true value.

It might also happen that the ToF is negative because the devices
involved in the CS RTT measurements are not perfectly calibrated for
internal delays.

In these cases it makes more sense to just set the estimated distance to
0.0f, since it is the best estimate we can make, given that we do not
want to use negative distances.

Signed-off-by: Erik Sandgren <erik.sandgren@nordicsemi.no>