ci: oss-history: disable for Mbed TLS
The history has been rewritten to start from a clean upstream state
in the Zephyr fork, so now the script tries to cherry pick upstream
Mbed TLS commits because they have different hashes between the
repos and then fails because the change is actually already present,
it's just that the histories have diverged.

On top of that the oss-history check will always use the manifest
revision from the merge base for zephyr and sdk-zephyr (upmerge),
so even making a PR to sdk-zephyr won't help.

Doesn't seem worth the trouble so disable the check for Mbed TLS.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>