lib: lte_link_control: Apply missing settings after +CFUN=0
Band lock, RAI notification subscription and DNS fallback
address were not applied after modem was put into minimal
functionality mode (AT+CFUN=0). Those settings are not
stored into NVM by the modem and the settings need to be
applied again.

Corrected the comments for PLMN lock in on_modem_init().
That setting is not volatile, but it is saved into NVM
by the modem.

Changed on_modem_init() callback to always continue execution
in case setting a configuration to modem fails. Earlier some of
the configurations caused the function to exit, while some did
not. The callback does not return any value, so it does its
best and tries to configure everything. Possible failures are
logged.

Jira: NRF91-2680

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>