Kconfig: Set default stack size for HW_UNIQUE_KEY_RANDOM
Set the default stack size to be 2048 for all the users
of the HW_UNIQUE_KEY_RANDOM option. This option requires
stack size >= 2048 so this makes sense to be the default.

Until now there were users of this option which got stack
size set to 2048 because they happened to enabled
ENTROPY_CC3XX but now that more users switch to
PSA crypto this entropy option will not be enabled.
But in general it makes more sense to add the default
based on the actual requirement.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>