drivers: flash_sync_mpsl: Reduce timeout for low priority attempt
This commit reduces the time it takes to elevate the priority
of a flash operation from 30 ms to 10 ms.

In applications where the radio is heavily used, this improves
the responsiveness. Examples of such application types:
 - Bluetooth applications using relatively low connection intervals
 - Bluetooth applications perfoming background scanning
 - 802.15.4 applications
 - Bluetooth LE audio applications

For applications where this is not the case, it is unlikely that
any behavior change will be seen.

This change also prepares the driver for fixing a bug in the scheduler
where the timeslot request timed out before the specified timeout.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>