Bluetooth: Controller: Select multirole library if MPSL_FEM is enabled
To reduce the code size for the optimized Peripheral and Central
libraries, the multirole library must be used if FEM support is wanted.
This is in preparation for FEM to be removed from the Central and
Peripheral libraries.

MPSL_CX is only supported in the multirole library.
Add guards in case someone explicitly selects the Peripheral or Central
library with the unsupported functionality.

Signed-off-by: Vegard Gulaker <vegard.gulaker@nordicsemi.no>