nrf_security: drivers: cracen: unlock mutex on gen pub key failure
In ecc_genpubkey; there are return paths that do not unlock the
asymmetric mutex locked in sx_pk_acquire_req() which may lead to
indefinite blocked code execution. This commit attempts to fix that by
setting the return value and breaking out of the while loop to return
that value after calling sx_pk_release_req() which unlocks the mutex.

This also revives the dead code which increments the attempts counter,
as the function would always return before that point if status is
SX_ERR_NOT_INVERTIBLE.

Also initialized the outputs pointer to NULL to avoid maybe-unitialized
compiler warnings.

Signed-off-by: Theis Mejnertsen <theismejnertsen@gmail.com>