sysbuild: Adapt build system to support nRF54H radio code reloc to TCM
Adapt the build system to support the nRF54H radio code relocation
from NVM to TCM RAM using the CONFIG_XIP=n option.
The CMake script of the variant image (secondary slot image for
images compatible with MCUboot direct-xip mode) needs to be modified
to properly pass correct addresses to the CONFIG_BUILD_OUTPUT_ADJUST_LMA
Kconfig option. It is a workaround and needs to be aligned in the
future.

JIRA: NCSDK-36461

Signed-off-by: Jan Zyczkowski <jan.zyczkowski@nordicsemi.no>
Co-authored-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>