applications: nrf_desktop: dfu_mcumgr: align mcuboot modes
Updated the Kconfig options of the DFU MCUmgr application module that
is part of the nRF Desktop application.

This module cannot be enabled when the MCUboot RAM load mode is used.

The CONFIG_DESKTOP_DFU_MCUMGR_MCUBOOT_DIRECT_XIP Kconfig option was
also removed to simplify maintainence. The module uses the
CONFIG_MCUBOOT_BOOTLOADER_MODE_DIRECT_XIP Kconfig option to identify
the direct-xip mode. This option is directly supplied by the MCUboot
component.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>