applications: nrf_desktop: add constraints for MCUboot RAM load
Added the DTS RAM layout constraints to the nRF Desktop application
when it uses the MCUboot bootloader in the RAM load mode. These
constraints enforce a specific RAM layout for supported board targets.
The user is required to define two DTS nodes that are part of the
chosen SRAM DTS node of the application image:

- cpuapp_sram_app_rxm_region
- cpuapp_sram_mcuboot_ram_region

Additionally, the cpuapp_sram_app_rxm_region DTS node must be placed
before the cpuapp_sram_mcuboot_ram_region DTS node in the target
device memory.

Added a linker script with the assert statement that prevents the
executable RAM region of the application image to overflow the
standard RAM region of the MCUboot image.

Ref: NCSDK-35506

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>