doc: ironside: Consolidate release archive info
Programming steps are discussed in the update docs, as that process
is either part of the manual update or provisioning; the latter
being documented elsewhere.

Describing the ZIP archive is expanded into a release deliverable
section and moved into the update context.

Default configurations aren't strictly related to either but are still
important to discuss in general ISE.

Ref: NCSDK-33336

Signed-off-by: Stephen Stauts <stephen.stauts@nordicsemi.no>