nrf_security: CRACEN: Fix issue with hanging IKG on CRACEN lite
Fix an issue with the IKG hanging in some /ns builds.
This was due to a previously fixed issue with entropy errors
in CRACEN lite due to misconfiguration of the TRNG.
This was solved by updating the repeat threshold and
propthreshold values, but this only happened when
the TRNG was opened, which mean it would still fail
silently for IKG operations.

Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>
(cherry picked from commit f83904b660f21dd5d4b4c12b0ad7e2f39ee2827b)