tests: support swapping UART
Demonstrate how to update the PERIPHCONF blob pointed to by
UICR.PERIPHCONF in a system where Direct XIP is used.

The application firmware is executed from one of two slots when Direct
XIP is in use, this means that the global PERIPHCONF blob contained in
the periphconf partition needs to be updated by the application firmware
if it has to be updated.

Firstly, we integrate the PERIPHCONF blob of each image within the
firmware of the image. This way its contents is covered by the signature
validation performed by MCUBoot, and any updates to the PERIPHCONF is
contained within the firmware itself.

Next, we add a function that is executed in SYS_INIT(EARLY), that is
before any device driver setup is done, which ensures that the global
PERIPHCONF matches the local one contained in the image firmware. If it
does not match, the global PERIPHCONF is updated from the point of the
first mismatch.

We also add a mechanism showing how to re-order the PERIPHCONF so that
the entries from MCUBoot are first. This is done to ensure that MCUBoot
is not affected by any updates to the global PERIPHCONF, and that the
system is power-failure safe during an update to the global PERIPHCONF.

Some comments on the changes introduced in this commit:
- The swap snippet simply swaps which UART is used for an image
- We re-use an existing application
  (tests/subsys/bootloader/upgrade/ref_smp_svr/CMakeLists.txt)
  and create a CMake parameter 'PERIPHCONF_BIN' that is used to pass
  local PERIPHCONF to the image firmware. When this is set, the lib
  for copying the local PERIPHCONF to the global one is included.
- The copy lib does not use any driver as that code is not available
  at the time of execution (SYS_INIT(EARLY)).
- Add build configurations for this application to make build
  commands more understandable.
- The rest of the commit are helper scripts to create the new
  PERIPHCONF blob and verify the functionality.

The steps to build and verify the changes are:

 # cd to the test application
cd $(west topdir)/nrf/tests/subsys/bootloader/upgrade/ref_smp_svr

 # Build the sample and generate all required artifacts
./move_mcuboot_periphconf_first.sh

 # Program the initial firmware
west flash --build-dir build_no_mcuboot_uart --recover

 # Program the updated periphconf which has MCUBoot entries first
nrfutil device program  --firmware build_no_mcuboot_uart/periphconf_final.hex \
--options chip_erase_mode=ERASE_NONE

 # Reset and verify that app works as expected
nrfutil device reset --reset-kind RESET_PIN --traits jlink

 # Close any UART connection at this point, otherwise next step will fail

 # Upload the update (takes some minutes)
mcumgr --conntype serial --connstring "/dev/ttyACM0,baud=115200" image upload \
build_swapped_uart_v2/zephyr_secondary_app.signed.bin

 # Open UART to see logs, then trigger a reset
nrfutil device reset --reset-kind RESET_PIN --traits jlink

Signed-off-by: Håkon Amundsen <haakon.amundsen@nordicsemi.no>