samples: bluetooth: fast_pair: locator_tag: fix CMake for UI module
Fixed the CMake description of the application UI module that is part
of the Fast Pair Locator Tag sample.

Changed the location of the public header file for the UI module.

Simplified the application module directory by removing the
include_priv/ directory. Renamed the app_ui_speaker.h private
header by appending the "priv" suffix to align it with other
private headers.

Ref: NCSDK-29877

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>