mpsl: clock_ctrl: Do not allow to use to short HFXO rampup by MPSL
The MPSL uses HFXO ramp-up time find instant to request the clock
before radio event. In the meantime, there are other operations
happening that must be completed before the radio event. In other
case, the event will be aborted. We can't accept too short ramp-up
time to make room for the low priority context to execute.

The value selected 850us is taken from current:
zephyr/boards/nordic/nrf54h20dk/bicr.json

That is safe for us, because we use it in our CI.

NOTE: the minimum time for HFXO ramp-up must be secured for 54h.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>