samples: ipc_service: Enable cpurad on nRF54H20 in Kconfig overlay
Enable the `SOC_NRF54H20_CPURAD_ENABLE` Kconfig option in the overlay
for the target board instead of using for this the `extra_args` property
in a specific test item defined in `sample.yaml`. This way it is more
visible that this option needs to be enabled and additionally the sample
can also be used successfully when built without supplying any test item
to `west build`.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>