doc: Add a known issue entry regarding Matter Bridge
While adding a dynamic endpoint to the Matter Bridge over Thread
application, it does not appear in commercial Matter ecosystem
applications.

This issue occurs because the Matter stack thread is not locked
during the addition or removal of a dynamic endpoint. As a result,
a race condition can happen when the
`CONFIG_CHIP_USE_ZEPHYR_NETWORKING` option is set to ``n``.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>