bluetooth: services: make cs de independent of RAS
up until now the only reasonable way to calculate distance with BLE CS
was through use of the RAS service. However, with the introduction of
the inline PCT transfer (IPT) feature, it is possbile for an initiator
to calculate distance without ever receiving measurements from the
reflector.

In order to facilitiate this change the distance estimation parts of the
CS DE library have been made independent of RAS so as to allow for
easier integration of IPT in the future.

Signed-off-by: Thomas Johansen <thomas.johansen@nordicsemi.no>